move dcmread wrapper into separate file, to prevent circular import
