+++ deid/tests/test_action_interaction.py
@@ -9,7 +9,7 @@ import shutil
 import tempfile
 import unittest
 
-from deid.dicom.utils import dcmread
+from deid.dicom.dcmread import dcmread
 
 from deid.data import get_dataset
 from deid.dicom import replace_identifiers
