+++ deid/tests/test_blank_action.py
@@ -9,7 +9,7 @@ import shutil
 import tempfile
 import unittest
 
-from deid.dicom.utils import dcmread
+from deid.dicom.dcmread import dcmread
 
 from deid.data import get_dataset
 from deid.dicom import replace_identifiers
