+++ deid/tests/test_clean_pixel_dimensions.py
@@ -10,7 +10,7 @@ import shutil
 import tempfile
 import unittest
 
-from deid.dicom.utils import dcmread
+from deid.dicom.dcmread import dcmread
 from deid.data import get_dataset
 from deid.utils import get_installdir
