+++ deid/tests/test_replace_action.py
@@ -10,7 +10,7 @@ import tempfile
 import unittest
 
 
-from deid.dicom.utils import dcmread
+from deid.dicom.dcmread import dcmread
 from deid.data import get_dataset
 from deid.dicom import replace_identifiers
 from deid.tests.common import create_recipe, get_file
