+++ deid/tests/test_sequence_jitter.py
@@ -9,7 +9,7 @@ import shutil
 import tempfile
 import unittest
 
-from deid.dicom.utils import dcmread
+from deid.dicom.dcmread import dcmread
 from deid.data import get_dataset
 from deid.dicom import get_files, replace_identifiers
 from deid.tests.common import create_recipe
